Renamed the form view page component. It was just named "page", and renamed the component to "FormViewContent".
