+++ src/app/form-view/[id]/page.tsx
@@ -2,11 +2,11 @@
 import Link from "next/link"
 import { useParams } from "next/navigation"
 import LoadingSpinner from "~/app/_components/LoadingSpinner"
-import FormView from "~/app/_components/form/FormView"
+import FormViewContent from "~/app/_components/form/FormViewContent"
 import { api } from "~/trpc/react"
-import styles from "../../_components/form/FormView.module.scss"
+import styles from "../../_components/form/FormViewContent.module.scss"
 
-export default function page() {
+function FormView() {
   const params = useParams<{ id: string }>()
   const { data, isLoading } = api.form.getPublicFormById.useQuery(params.id)
 
@@ -43,5 +43,7 @@ export default function page() {
     )
   }
 
-  return <div>{data && <FormView formData={data} />}</div>
+  return <div>{data && <FormViewContent formData={data} />}</div>
 }
+
+export default FormView
